created template directory for experiments, refactored liquid_solid_extraction and tested
