+++ escalate/core/experiment_templates/__init__.py
@@ -1,0 +1,1 @@
+from .liquid_solid_extraction import *
